try-except

+++ PSAAI/SubjectList/views.py
@@ -1,5 +1,7 @@
 from datetime import date, timedelta
 
+from django.contrib.auth.mixins import LoginRequiredMixin
+from django.db import IntegrityError, DatabaseError
 from django.db.models import Count
 from django.http import HttpResponse
 from django.shortcuts import render, redirect
@@ -11,15 +13,19 @@ from .models import *
 from django.views.generic import TemplateView
 
 
-class Academia(TemplateView):
+class Academia(LoginRequiredMixin, TemplateView):
     template_name = 'SubjectList/academia.html'
 
     def get_context_data(self, *args, **kwargs):
         context = super(Academia, self).get_context_data(**kwargs)
-        context['sciences'] = Course.objects.filter(discipline='science')
-        context['literature'] = Course.objects.filter(discipline='literature')
-        context['social'] = Course.objects.filter(discipline='social')
 
+        try:
+            context['sciences'] = Course.objects.filter(discipline='science')
+            context['literature'] = Course.objects.filter(discipline='literature')
+            context['social'] = Course.objects.filter(discipline='social')
+
+        except DatabaseError as error:
+            context['error_message'] = "Failed to retrieve data from the database. Please try again later."
         return context
 
     def post(self, request, **kwargs):
@@ -28,21 +34,21 @@ class Academia(TemplateView):
             subjects = request.POST.getlist('subjects')
             user = self.request.user
             try:
-                my_subjects = MySubjects.objects.get(user=user)
+                my_subjects = MySubjects.objects.filter(user=user).first()
                 my_subjects.name.set(subjects)
                 my_subjects.save()
+
                 return redirect('student-home')
 
-            except MySubjects.DoesNotExist:
-                return HttpResponse(
-                    'DatabaseError: We could not save your subject Selection')
+            except DatabaseError as error:
+                return redirect('student-home')
 
 
-class Learning(TemplateView):
+class Learning(LoginRequiredMixin, TemplateView):
     template_name = 'SubjectList/select_subject.html'
 
 
-class Read(TemplateView):
+class Read(LoginRequiredMixin, TemplateView):
     template_name = 'SubjectList/read.html'
 
     def get_context_data(self, **kwargs):
@@ -50,103 +56,109 @@ class Read(TemplateView):
         name = self.kwargs['name']
 
         try:
-            context['subject'] = Subtopic.objects.get(name=name)
+            context['subject'] = Subtopic.objects.filter(name=name).first()
             return context
 
-        except Subtopic.DoesNotExist:
-            return HttpResponse('Subtopic does not exist.')
-
-        except Subtopic.DoesNotExist:
-            return HttpResponse('Subtopic does not exist.')
-
-        except Subtopic.MultipleObjectsReturned:
-            return HttpResponse('More than one Subtopic returned.')
+        except DatabaseError as error:
+            pass
 
 
 
-
-class Finish(TemplateView):
+class Finish(LoginRequiredMixin, TemplateView):
     template_name = 'SubjectList/finish.html'
 
     def get_context_data(self, **kwargs):
         context = super(Finish, self).get_context_data(**kwargs)
         try:
-            subtopic = Subtopic.objects.get(name=self.kwargs['name'])
+            subtopic = Subtopic.objects.filter(name=self.kwargs['name']).first()
             topic = subtopic.topic.name
             context['topic'] = topic
+
             return context
 
-        except Subtopic.DoesNotExist:
-            return HttpResponse('Subtopic does not exist.')
-
-        except Subtopic.DoesNotExist:
-            return HttpResponse('Subtopic does not exist.')
-
-        except Subtopic.MultipleObjectsReturned:
-            return HttpResponse('More than one Subtopic returned.')
-
-
-
+        except DatabaseError as error:
+            pass
 
     def post(self, request, **kwargs):
         if request.method == 'POST':
             user = request.user
             try:
-                subtopic = Subtopic.objects.get(name=self.kwargs['name'])
+                subtopic = Subtopic.objects.filter(name=self.kwargs['name']).first()
                 topic = subtopic.topic.name
-                topic = Topic.objects.get(name=topic)
-                subject = Subject.objects.get(name=subtopic.subject)
+                topic = Topic.objects.filter(name=topic).first()
+                subject = Subject.objects.filter(name=subtopic.subject).first()
                 about = f'{subject}: {topic} quiz is ready.'
                 message = 'The quiz for this topic is now ready. Once started the quiz will finish in 15 minutes. Good luck.'
                 is_progress = Progress.objects.filter(user=self.request.user, topic=topic, subtopic=subtopic)
+
                 if is_progress.exists():
                     pass
                 else:
-                    progress = Progress.objects.create(user=user, subtopic=subtopic, subject=subject)
-                    progress.topic.set([topic])
-                    progress.save()
+                    try:
+                        progress = Progress.objects.create(user=user, subtopic=subtopic, subject=subject)
+                        progress.topic.set([topic])
+                        progress.save()
+                    except IntegrityError:
+                        pass
                     total_topics = topic.topics_count
-                    all_subtopics = Progress.objects.filter(user=user, topic=topic).values('subtopic').distinct().count()
+                    all_subtopics = Progress.objects.filter(user=user, topic=topic).values(
+                        'subtopic').distinct().count()
                     print(all_subtopics)
                     if all_subtopics == int(total_topics):
-                        notification = TopicExamNotifications.objects.create(user=user, about=about, message=message,
-                                                                             topic=topic)
+                        try:
+                            notification = TopicExamNotifications.objects.create(user=user, about=about,
+                                                                                 notification_type='quiz',
+                                                                                 subject=subject, message=message,
+                                                                                 topic=topic)
+                        except IntegrityError as error:
+                            return HttpResponse(f'{error},We could not create a test for you . Please contact Admin.')
                     else:
                         pass
-            except subtopic.MultipleObjectsReturned:
-                return HttpResponse('We could not save your progress, try again.')
+
+            except DatabaseError as error:
+                pass
+
+            except Exception as error:
+                return HttpResponse(f'{error}')
 
         return redirect('student-home')
 
 
-class Syllabus(TemplateView):
+class Syllabus(LoginRequiredMixin, TemplateView):
     template_name = 'SubjectList/syllabus.html'
 
     def get_context_data(self, **kwargs):
         context = super(Syllabus, self).get_context_data(**kwargs)
         subject = self.kwargs['name']
-        context['syllabus'] = Topic.objects.filter(subject__name=subject).order_by('order')
+        try:
+            context['syllabus'] = Topic.objects.filter(subject__name=subject).order_by('order')
+
+        except DatabaseError as error:
+            pass
 
         return context
 
 
-class Messages(TemplateView):
+class Messages(LoginRequiredMixin, TemplateView):
     template_name = 'SubjectList/messages.html'
 
     def get_context_data(self, **kwargs):
         context = super(Messages, self).get_context_data(**kwargs)
         user = self.request.user
-        topical_exam_results = TopicalExamResults.objects.filter(user=user)
-        topical_exam = TopicExamNotifications.objects.filter(user=user)
-        class_bookings = ClassBookingNotifications.objects.filter(user=user)
-        subscription_notifications = SubscriptionNotifications.objects.filter(user=user)
-        payment_notification = PaymentNotifications.objects.filter(user=user)
+        try:
+            topical_exam_results = TopicalExamResults.objects.filter(user=user)
+            topical_exam = TopicExamNotifications.objects.filter(user=user)
+            class_bookings = ClassBookingNotifications.objects.filter(user=user)
+            subscription_notifications = SubscriptionNotifications.objects.filter(user=user)
+            payment_notification = PaymentNotifications.objects.filter(user=user)
 
-        messages = list(topical_exam) + list(topical_exam_results) + list(class_bookings) + list(
-            subscription_notifications) + list(payment_notification)
-        context['messages'] = messages
+            messages = list(topical_exam) + list(topical_exam_results) + list(class_bookings) + list(
+                subscription_notifications) + list(payment_notification)
+            context['messages'] = messages
+            return context
 
-        return context
+        except DatabaseError as error:
+            pass
 
 
 class MyProgress(TemplateView):
@@ -154,16 +166,19 @@ class MyProgress(TemplateView):
 
     def get_context_data(self, **kwargs):
         context = super(MyProgress, self).get_context_data(**kwargs)
-        subject = Progress.objects.filter(user=self.request.user, subject__isnull=False).values('subject__name',
-                                                                                                'subject__topics').annotate(
-            topic_count=Count('topic', distinct=True))
-        count = Progress.objects.filter()
+        try:
+            subject = Progress.objects.filter(user=self.request.user,
+                                              subject__isnull=False).values('subject__name', 'subject__topics'). \
+                annotate(topic_count=Count('topic', distinct=True))
 
-        context['subject'] = subject
-        return context
+            context['subject'] = subject
+            return context
+
+        except DatabaseError as error:
+            pass
 
 
-class UpcomingClasses(TemplateView):
+class UpcomingClasses(LoginRequiredMixin, TemplateView):
     template_name = 'SubjectList/upcoming_classes.html'
 
     def get_context_data(self, **kwargs):
@@ -174,67 +189,93 @@ class UpcomingClasses(TemplateView):
         end_date = today + timedelta(days=7)
 
         # Filter OnlineClass objects with a date within the
-        # OnlineClass.objects.filter(date__range=(today, end_date))
-        upcoming_classes = OnlineClass.objects.all()
+        try:
+            # OnlineClass.objects.filter(date__range=(today, end_date))
+            upcoming_classes = OnlineClass.objects.all()
 
-        context['classes'] = upcoming_classes
-        return context
+            context['classes'] = upcoming_classes
+            return context
+
+        except DatabaseError as error:
+            pass
 
 
-class ClassBookings(TemplateView):
+class ClassBookings(LoginRequiredMixin, TemplateView):
     template_name = 'SubjectList/class_booking.html'
 
     def get_context_data(self, **kwargs):
         context = super(ClassBookings, self).get_context_data(**kwargs)
         class_id = self.kwargs['id']
-        work = OnlineClass.objects.get(id=class_id)
-        context['class'] = work
 
-        return context
+        try:
+            work = OnlineClass.objects.filter(id=class_id).last()
+            context['class'] = work
+            return context
+
+        except DatabaseError as error:
+            pass
 
     def post(self, request, **kwargs):
         user = request.user
         class_id = self.kwargs['id']
-        class_instance = OnlineClass.objects.get(id=class_id)
-        action = request.POST.get('action')
-        if 'book-class' in action:
+        try:
 
-            booking = ClassBooking.objects.create(user=user, class_name=class_instance)
-            return redirect('student-home')
-        elif 'delete-class' in action:
-            booking = ClassBooking.objects.filter(user=user, class_name=class_instance).first()
-            print(booking)
-            if booking:
-                booking.delete()
-            return redirect('student-home')
-        else:
-            return redirect('notifications')
+            class_instance = OnlineClass.objects.filter(id=class_id).last()
+            action = request.POST.get('action')
+            if class_instance:
+                if 'book-class' in action:
+
+                    try:
+                        booking = ClassBooking.objects.create(user=user, class_name=class_instance)
+                        return redirect('student-home')
+                    except IntegrityError:
+                        pass
+                elif 'delete-class' in action:
+                    booking = ClassBooking.objects.filter(user=user, class_name=class_instance).first()
+                    print(booking)
+                    if booking:
+                        booking.delete()
+                    return redirect('student-home')
+                else:
+                    return redirect('notifications')
+
+        except DatabaseError as error:
+            pass
 
 
-class BookedClasses(TemplateView):
+class BookedClasses(LoginRequiredMixin, TemplateView):
     template_name = 'SubjectList/booked_classes.html'
 
     def get_context_data(self, **kwargs):
         context = super(BookedClasses, self).get_context_data(**kwargs)
-        bookings = ClassBooking.objects.filter(user=self.request.user)
-        context['bookings'] = bookings
 
-        return context
+        try:
+            bookings = ClassBooking.objects.filter(user=self.request.user)
+            context['bookings'] = bookings
+
+            return context
+
+        except DatabaseError as error:
+            pass
 
 
-class CallLobby(TemplateView):
+class CallLobby(LoginRequiredMixin, TemplateView):
     template_name = 'SubjectList/lobby.html'
 
     def get_context_data(self, **kwargs):
         context = super(CallLobby, self).get_context_data(**kwargs)
         class_id = self.kwargs['id']
-        class_instance = VideoChannel.objects.get(class_id=class_id)
-        context['class'] = class_instance
+        try:
+            class_instance = VideoChannel.objects.filter(class_id=class_id).last()
+            context['class'] = class_instance
 
-        return context
+            return context
+
+        except DatabaseError as error:
+            pass
 
 
-class VideoCall(TemplateView):
+class VideoCall(LoginRequiredMixin, TemplateView):
     template_name = 'SubjectList/video_call.html'
 
     def get_context_data(self, **kwargs):
@@ -243,3 +284,54 @@ class VideoCall(TemplateView):
         context['app_id'] = app_id
 
         return context
+
+
+class ContactUs(LoginRequiredMixin, TemplateView):
+    template_name = 'SubjectList/contact.html'
+
+    def get_context_data(self, **kwargs):
+        context = super(ContactUs, self).get_context_data(**kwargs)
+
+        try:
+            subject = Course.objects.all()
+            context['subject'] = subject
+
+            return context
+
+        except DatabaseError as error:
+            pass
+
+    def post(self, request, **kwargs):
+
+        if request.method == 'POST':
+            user = request.user
+            message = request.POST.get('message')
+            about = request.POST.get('about')
+
+            if about == 'Academic':
+                subject = request.POST.get('subject')
+
+                try:
+                    subject = Subject.objects.filter(name=subject).first()
+                    try:
+
+                        inquiry = AcademicInquiries.objects.create(user=user, quiz_class=about, subject=subject,
+                                                               message=message)
+                        return redirect('student-home')
+
+                    except IntegrityError:
+                        pass
+
+                except DatabaseError as error:
+                    pass
+
+            elif about == 'Account':
+                try:
+                    inquiry = AccountInquiries.objects.create(user=user, quiz_class=about, message=message)
+
+                    return redirect('student-home')
+
+                except IntegrityError:
+                    pass
+            else:
+                return HttpResponse('could not send your inquiry. Please try again')
